app/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.1
psycopg2-binary==2.9.9
pycparser==2.21
pyOpenSSL==23.2.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
typing_extensions==4.8.0
Werkzeug==2.3.7
